+++ app/coins/btcHolidays.js
@@ -1,0 +1,115 @@
+module.exports = {
+	desc: "A curated list of unofficial Bitcoin 'Holidays'. Adapted from https://bitcoin.holiday, and elsewhere.",
+	items: [
+		{
+			name: "Bitcoin Pizza Day",
+			date: "2010-05-22",
+			desc: "On this day in 2010, Lazlo Hanyecz, paid Jeremy Sturdivant 10,000 BTC for two pizzas costing ~41 USD. The famous <a href='./tx/a1075db55d416d3ca199f55b6084e2115b9345e16c5cf302fc80e9d5fbf5d48d@57043'>\"Bitcoin Pizza\" transaction</a> was one of the earliest real-world BTC transactions and is celebrated as helping start the Bitcoin economy."
+		},
+		{
+			name: "Shitcoin Day",
+			date: "2013-07-03",
+			desc: "This day marks the beginning of the first ever \"utility token\" ICO, called Mastercoin (later rebranded as Omni)."
+		},
+		{
+			name: "SegWit Lock-In Day",
+			date: "2017-08-08",
+			desc: "Segregated Witness is officially locked in by the UASF client on the Bitcoin network."
+		},
+		{
+			name: "Bitcoin Infinity Day",
+			date: "2021-08-21",
+			desc: "8/21 is playfully celebrated as representing the meme \"Everything there is divided by 21,000,000\" (∞/21M)."
+		},
+		{
+			name: "SegWit Activation Day",
+			date: "2017-08-23",
+			desc: "On this day in 2017, at 1:57 UTC, SegWit activated on the Bitcoin mainnet in <a href='./block-height/481824'>Block #481,824</a>."
+		},
+		{
+			name: "Bitcoin Legal Tender Day",
+			date: "2021-09-07",
+			desc: "On this day in 2021, BTC became legal tender in El Salvador, 90 days after the publication of the law in the official gazette, making El Salvador the first country to enact BTC as legal tender through legistlation."
+		},
+		{
+			name: "Free Ross Day",
+			date: "2013-10-01",
+			desc: "On this day in 2013, Ross Ulbricht was captured and imprisoned for his role in creating The Silk Road website. He was given two life sentences for non-violent crimes. Fighting for Ross's freedom is a cause supported by many Bitcoiners."
+		},
+		{
+			name: "BearWhale Slay Day",
+			date: "2014-10-06",
+			desc: "On this day in 2014, a bitcoin whale posted a 30,000 BTC sell wall, resulting in an epic skirmish between hodlers and the BearWhale. Over the course of six hours, the price stalled at 300 USD until all of the coins sold off and the BearWhale was defeated. You can watch the battle <a href='https://twitter.com/DocumentingBTC/status/1364230687105167362?s=20The'>here</a>."
+		},
+		{
+			name: "This is Gentlemen! Day",
+			date: "2014-11-11",
+			desc: "On this day in 2014, an over-excited Reddit user accidentally omitted a word and unknowingly earned a bizarre place in Bitcoin culture and lore. Read the Reddit daily discussion on Bitcoin Markets <a href='https://www.reddit.com/r/BitcoinMarkets/comments/2lxkqk/daily_discussion_tuesday_november_11_2014/'>here</a>."
+		},
+		{
+			name: "HODL Day",
+			date: "2013-12-18",
+			desc: "On this day in 2013, a user on the bitcointalk.org declared <a href='https://bitcointalk.org/index.php?topic=375643.0'>I AM HODLING</a> (!), rather than selling, despite the falling price. This \"whisky\"-facilitated post spawned one of Bitcoin's great memes."
+		},
+		{
+			name: "Bitcoin's Birthday",
+			date: "2009-01-03",
+			desc: "On this day in 2009, the <a href='./block-height/0'>Genesis Block</a> was mined by Satoshi Nakamoto, officially starting the Bitcoin blockchain."
+		},
+		{
+			name: "Proof of Keys Day",
+			date: "2019-01-03",
+			desc: "On this day in 2019, to celebrate Bitcoin's 10th Birthday, Trace Mayer started an annual Proof-of-Keys Celebration. Participats declare monetary sovereignty by withdrawing their BTC into their own custory. Not your keys, not your bitcoin! You can read Trace Mayer's original tweet <a href='https://twitter.com/TraceMayer/status/1071870548421066753'>here</a>."
+		},
+		{
+			name: "Genesis Transaction Day",
+			date: "2009-01-11",
+			desc: "On this day in 2009, the first transaction ever broadcast over the Bitcoin network was sent by Satoshi Nakamoto to Hal Finney. View the genesis transaction <a href='./tx/f4184fc596403b9d638783cf57adfe4c75c605f6356fbc91338530e9831e9e16@170'>here</a>."
+		},
+		{
+			name: "Lightning Pizza Day",
+			date: "2018-02-25",
+			desc: "Eight years after buying pizza with the famous <a href='./tx/a1075db55d416d3ca199f55b6084e2115b9345e16c5cf302fc80e9d5fbf5d48d@57043'>\"Bitcoin Pizza\" transaction</a>, Lazlo Hanyecz returned to buy pizza over the Lightning Network."
+		},
+		{
+			name: "Mt. Gox Death Day",
+			date: "2013-02-24",
+			desc: "On this day in 2013, Mt. Gox, the largest Bitcoin exchange in the world, went offline due to insolvency, after losing 640,000 BTC."
+		},
+		{
+			name: "Gold Parity Day",
+			date: "2017-03-03",
+			desc: "On this day in 2017, 1 BTC surpassed the price of 1 ounce of gold for the first time."
+		},
+		{
+			name: "Billion Dollar Day",
+			date: "2013-03-28",
+			desc: "On this day in 2013, the market cap of all BTC surpassed 1 billion USD for the first time."
+		},
+		{
+			name: "Trillion Dollar Day",
+			date: "2021-02-19",
+			desc: "On this day in 2021, the market cap of all BTC surpassed 1 trillion USD for the first time, ~8 years after surpassing 1 billion USD for the first time."
+		},
+		{
+			name: "Bitcoin ATM Day",
+			date: "2013-05-02",
+			desc: "On this day in 2013, the first-ever Bitcoin ATM was installed at a Vancouver coffee shop. It included a built-in palm scanner designed to prevent users from processing more than 3,000 CAD per day."
+		},
+		{
+			name: "Bitcoin Whitepaper Day",
+			date: "2008-10-31",
+			desc: "On this day in 2008, Satoshi Nakamoto released the Bitcoin white paper to the cryptography mailing list, marking its official introduction into the world. Satoshi introduced Bitcoin as an electronic cash system with several important properties: double-spending prevention, no trusted parties, anonymous participation, and Hashcash style Proof-of-Work. The Bitcoin Whitepaper can be viewed <a href='./bitcoin-whitepaper'>here</a>. "
+		},
+		{
+			name: "Bitcoin Logo Day",
+			date: "2011-11-04",
+			desc: "On this day in 2010, an unknown artist under the username \"bitboy\" offered the Bitcoin community its iconic logo for free on the Bicointalk forum. You can view the original post <a href='https://bitcointalk.org/index.php?topic=1631.0%3Ball'>here</a>."
+		},
+		{
+			name: "Diffe-Hellman Day",
+			date: "1976-11-01",
+			desc: "On this day in 1976, Whitfield Diffie and Martin E. Hellman released their proposal for a breakthrough in cryptographic communication: the first implementation of public/private key pairs for secure communication over public channels. This cryptographic building block is one of the key underpinnings of the Bitcoin network. Read their 1976 paper titled \"New Directions in Cryptography\" <a href='https://ee.stanford.edu/~hellman/publications/24.pdf'>here</a>."
+		}
+	]
+};